Make a ggplot2 model diagram.

Make a ggplot2 model diagram.

make_diagram(
  df_list,
  label_flows = TRUE,
  external_flows = TRUE,
  interaction_label = TRUE,
  with_grid = FALSE,
  node_outline_color = "black",
  node_fill_color = "white",
  node_text_color = "black",
  node_text_size = 8,
  flow_text_color = "black",
  flow_text_size = 3,
  main_arrow_color = "black",
  main_arrow_linetype = 1,
  main_arrow_size = 0.5,
  interaction_arrow_color = "black",
  interaction_arrow_linetype = 2,
  interaction_arrow_size = 0.5
)

Arguments

df_list
A list of data frames returned from the prepare_diagram function. See that function for details about this object.

label_flows
A logical indicating whether to label the flows (TRUE, default) or not (FALSE).

external_flows
A logical indicating whether to include flows into and out of the system (external flows). Default is TRUE (include).

interaction_label
A logical indicating whether to make the diagram with interaction terms (typically curved arrows leading to the mid point of another arrow) or to simply label the main flow. See vignettes for examples.

with_grid
A logical indicating whether to return the ggplot with a grid. Default is FALSE. The grid can be helpful if you want/need to move items around.

node_outline_color
A character string or vector of character strings specifying the color of node outlines. If a vector, the colors will be recycled in the order of the variables in the supplied data frame.

node_fill_color
A character string or vector of character strings specifying the fill color of nodes. If a vector, the colors will be recycled in the order of the variables in the supplied data frame.

node_text_color
A character string or vector of character strings specifying the text color for node labels. If a vector, the colors will be recycled in the order of the variables in the supplied data frame.

node_text_size
A numeric scalar specifying the text size for node labels. Default value is 8.

flow_text_color
A character string or vector of character strings specifying the text color for flow labels. If a vector, the colors will be recycled in the order of the flows in the supplied data frame.

flow_text_size
A numeric scalar specifying the text size for flow labels. Default value is 3.

main_arrow_color
A character string or vector of character strings specifying the text color for non-interaction flow arrows. If a vector, the colors will be recycled in the order of the flows in the supplied data frame.

main_arrow_linetype
A numeric scaler specifying the linetype for main arrows (non-interaction arrows).

main_arrow_size
A numeric scaler specifying the line size for the main arrows (non-interaction arrows).

interaction_arrow_color
A character string or vector of character strings specifying the text color for interaction flow arrows. If a vector, the colors will be recycled in the order of the flows in the supplied data frame.

interaction_arrow_linetype
A numeric scalar specifying the linetype for interaction arrows.

interaction_arrow_size
A numeric scalar specifying the line size for the interaction arrows.

Value

A ggplot2 object.