Create data frames for plotting from model elements.

Create data frames for plotting from model elements.

prepare_diagram(input_list, nodes_df = NULL)

Arguments

input_list
A list of model elements. The list must contain at least two elements with names varlabels and flows. The flows list must contain a sub-list for each variable in the varlabels vector. For example, if the user specifies two variables in input_list$varlabels, then input_list$flows must contain two sub-lists, each containing a character vector of flows into and out of the node. Currently, this function assumes that the input_list$varlabels sub-lists are in the same order as the input_list$varlabels vector. See examples. The input_list can contain any other elements that the user might deem useful (e.g., metadata/comments), but only the varlabels and flows are used by this function.

nodes_df
A data frame with user-specified node locations. The data frame must contain the following columns: id, label, x, and y. An internal function will add the necessary row column based on the values for y. See vignettes for examples of the data frame structure.

Value

A list of five data frames:

nodes: A data frame containing the node (state variable) ids, labels, and positions. Position is defined with the x and y columns, which define the center of the node squares. The ggplot2 function geom_tile is used for placement of nodes and the height and width are set to 1 to generate a perfect square centered on x and y. The row column indicates which row the node is placed on; this is also evident from the y values. id is a numeric id for the node, which shows up in the edge data frames for defining connects (see below).

horizontal_edges: A data frame containing name and position information for straight flows from one node to an adjacent node. The data frame contains nine (9) columns:

to: The node id to which the arrow will point. That is, the node receiving the flow.

from: The node id from which the arrow originate. That is, the node donating the flow.

label: The label of the flow. Typically a mathematical expression.

xstart: The starting horizontal position of the arrow.

ystart: The starting veritcal position of the arrow.

xend: The ending horizontal position of the arrow.

yend: The ending vertical position of the arrow.

xmid: The horizontal midpoint of the arrow. Used for label placement.

ymid: The vertical midpoint of the arrow. Used for label placement.

vertical_edges: A data frame containing name and position information for flows that arrive from out of the system or leave the system. That is, flows that either (1) do not come from a node specified in the nodes data frame or (2) exit the system without connecting to another node. The data frame contains the same columns as horizontal_edges. Note that either the to or from value will be NA in each row (if rows are present). These flows arrive or leave at 45 degree angles, despite the name of the data frame implying a vertical entrance or exit.

curved_edges: A data frame containing name and position information for, typically, two types of curved arrows: (1) interaction arrows that point toward a horizontal arrow or (2) a physical flow that would normally be a horizontal arrow but must bypass at least one node. The data frame has eleven (11) columns:

to: The node id to which the arrow will point. That is, the node receiving the flow.

from: The node id from which the arrow originate. That is, the node donating the flow.

label: The label of the flow. Typically a mathematical expression.

interaction: A logical indicating whether the flow represents an interaction between two or more nodes. If TRUE, the arrow is drawn as dashed by default.

xstart: The starting horizontal position of the arrow.

ystart: The starting veritcal position of the arrow.

xend: The ending horizontal position of the arrow.

yend: The ending vertical position of the arrow.

curvature: The amount of curvature applied to arrow. Higher numbers indicate more curvature; 0 = straight line.

row: The row on which the arrow is connecting nodes; this is also evident from the ystart and yend values.

labelx: Horizontal position of label.

labely: Vertical position of label.

feedback_edges: A data frame containing name and position information for arrows indicating a feedback into the same node. The data frame contains the same columns as the horizontal_edges data frame. Note that the to and from columns should have the same values for feedback edges.

Examples

varlabels <- c("S","I","R")
varnames <- c("Susceptible","Infected","Recovered")  # optional
flows <- list(S_flows = c("-b*S*I"),
              I_flows = c("+b*S*I","-g*I"),
              R_flows = c("+g*I"))
mymodel <- list(varlabels = varlabels, varnames = varnames, flows = flows)
prepare_diagram(input_list = mymodel)

#> $nodes
#>   id label row x  y
#> 1  1     S   1 3 -2
#> 2  2     I   1 6 -2
#> 3  3     R   1 9 -2
#> 
#> $horizontal_edges
#>   to from label xstart ystart xend yend xmid  ymid
#> 1  2    1            3     -2    6   -2  4.5 -1.75
#> 2  3    2   g*I      6     -2    9   -2  7.5 -1.75
#> 
#> $vertical_edges
#> [1] to     from   label  xstart ystart xend   yend   xmid   ymid  
#> <0 rows> (or 0-length row.names)
#> 
#> $curved_edges
#>   to from label interaction xstart ystart xend yend curvature row labelx labely
#> 1 NA    2 b*S*I        TRUE      6   -1.5  4.5   -2       0.4   1   5.15  -1.25
#> 
#> $feedback_edges
#> [1] to     from   label  xstart ystart xend   yend   xmid   ymid  
#> <0 rows> (or 0-length row.names)
#>