Create data frames for plotting from model elements.

This function takes as input a (typically) compartmental model consisting of variables/compartments and flows and creates a list of data frames with label and position information for plotting a flow diagram. The resulting object is used as an input to make_diagram, which creates a ggplot2 based diagram. Attempts to make decent decisions regarding the placement of nodes (boxes), flows (arrows), and labels are made. However, complex models with complex diagrams will likely need user modification. This is documented in the vignettes.

prepare_diagram(
  model_list,
  model_settings = list(varnames = NULL, use_varnames = FALSE, var_label_size = 10,
    varlocations = NULL)
)

Arguments

model_list
A list of model elements. This list is required and must contain these two elements:

varlabels: A character vector with labels for each variable.

flows: A list that is the same length as varlabels. Each sub-list element is a character vector of any length specifying the flows into and out of the variable. Note that flowdiagramr assumes that the order of flows and varlabels match.

IMPORTANT: All varlabel entries must start with an upper case letter, followed by any combination of letters and numbers (e.g., S, Si, or Aml2). All parameters contained in the flows must start with a lower case letter followed by any combination of letters and numbers (e.g., b, bBmax, kS, p21S). All variables and parameters MUST be separated by math notation (e.g., +, -, *, /).

See examples and details below and vignettes.

model_settings
A list of optional model settings. The following elements are supported and default values are provided:

varnames: Vector of strings containing labels, one for each variable. Default is NULL.

use_varnames: A logical. If FALSE (default) variable boxes will be labeled with the varlabels entry from model_list. If TRUE variable boxes will be labeled using the text provided in the varnames element. If varnames is missing, an error occurs. Note that labeling can be turned off through a setting in make_diagram.

var_label_size: A numeric defining the size of the variable labels in the plot. This is necessary because the the box sizes will (eventually) be determined by the size of the text within. Default is 10.

varlocations: A numeric matrix that specifies the locations of the variables on an x-y grid with their desired x (columns) and y (row) locations. See examples and vignettes. Default is NULL, which results in a left-to-right diagram.

Value

A list of two data frames:

variables: A data frame containing information for all variables. The data frame contains these columns:

label: The variable label as provided in the model specification.

name: If provided, the full text for each variable.

xmin: Left edge location of box.

xmax: Right edge location of box.

ymin: Lower edge of location box.

ymax: Upper edge of location box.

labelx: Horizontal position (midpoint) of label.

labely: Vertical position (midpoint) of label.

plot_label: The text to be written into the box.

color: Default outline color for the box.

fill: Default fill color for the box.

label color: Default color for text label.

label_size: Size of text to be written into the box.

flows: A data frame containing information for all flows. The data frame contains these columns:

to: The variable to which the arrow will point. That is, the variable receiving the flow.

from: The variable from which the arrow originate. That is, the variable donating the flow.

label: The label of the flow. Typically a mathematical expression.

xstart: The starting horizontal position of the arrow.

xend: The ending horizontal position of the arrow.

ystart: The starting vertical position of the arrow.

yend: The ending vertical position of the arrow.

labelx: Horizontal position (midpoint) of label.

labely: Vertical position (midpoint) of label.

curvature: The amount of curvature applied to arrow. Higher numbers indicate more curvature; 0 = straight line.

type: Type of flow. One of main, interaction, or external.

color: Default color of the lines/arrows.

linetype: Default linetype.

size: Default size of the lines.

label_color: Default label color.

label_size: Default text size for label.

arrowsize: Default arrow size.

math: The math from the flows specified by the user. Is a duplicate of label so that user can update label as desired but retain the original math for reference.

Details

varlabels needs to be specified as a vector of model variables, e.g., varlabels <- c("Pred","Prey"). flows need to be specified as a list, with each list entry containing the flows/processes for each variable in the order in which the variables appear. Flows need to be named according to VARLABEL_flows. Example: flows <- list(Pred_flows = c(r*Pred, -k1*Pred*Prey), Prey_flows = c(g*Prey, -k2*Pred*Prey) ) Each flow, i.e. each entry in the flow vector, needs to be a valid mathematical expression made up of varlabels and parameters. The rules are as described above. As an example, the following includes a parameter b and two variables, S and I: b*S*I. The following includes a parameter s and two variables, Bg and I2: Bg*s*I2. See more examples below and in the vignettes.

Examples

#basic model specification
varlabels <- c("S","I","R")
flows <- list(S_flows = c("-b*S*I"),
              I_flows = c("b*S*I","-g*I"),
              R_flows = c("g*I"))
mymodel <- list(varlabels = varlabels, flows = flows)
diag_list <- prepare_diagram(model_list = mymodel)
mydiag <- make_diagram(diag_list)

#adding optional specifications
varnames <- c("Susceptible","Infected","Recovered")
varlocations <-  matrix(data = c("S", "", "R",
                                 "", "I", "" ),
                        nrow = 2, ncol = 3, byrow = TRUE)
mysettings <- list(varnames = varnames, use_varnames = TRUE,
                   var_label_size = 4, varlocations = varlocations)
diag_list <- prepare_diagram(model_list = mymodel, model_settings = mysettings)
mydiag <- make_diagram(diag_list)

#another simple model
varlabels = c("Pat","Imm")
flows     = list(Pat_flows = c("g*Pat*(1-Pat/pmax)", "-dP*Pat", "-k*Pat*Imm"),
                 Imm_flows = c("r*Pat*Imm", "-dI*Imm"))
mymodel = list(varlabels = varlabels, flows = flows)
diag_list <- prepare_diagram(mymodel)
mydiag <- make_diagram(diag_list)

#options to switch to vertical layout and adding names
varnames <- c("Pathogen","Immune Response")
varlocations <-  matrix(data = c("Pat", "Imm"),
                        nrow = 2, byrow = TRUE)
mysettings <- list(varnames = varnames, use_varnames = TRUE,
                   var_label_size = 4, varlocations = varlocations)
diag_list <- prepare_diagram(mymodel,mysettings)
mydiag <- make_diagram(diag_list)