Generate a reproducible R script to make the diagram.

write_diagram() generates code in the form of a stand-alone R script to reproduce a diagram, including the model inputs, the data frames from prepare_diagram, and the ggplot2 code. The R script is intended to run "as-is", meaning there is code to make the objects the user sends as arguments (either the model or the diagram_list).

write_diagram(
  model_list = NULL,
  diagram_list = NULL,
  make_diagram_settings = NULL,
  use_varnames = FALSE,
  with_grid = FALSE,
  directory = NULL,
  filename = NULL
)

Arguments

model_list
A flowdiagramr input list. See prepare_diagram.

diagram_list
A flowdiagramr input structure, resulting from a call to prepare_diagram.

make_diagram_settings
A named list of diagram aesthetics. See make_diagram documentation. Default is NULL and the default values from make_diagram are used.

use_varnames
A logical indicating whether to label nodes with variable abbreviations (FALSE; default) or to use the full names provided in the varnames element of model_list (TRUE).

with_grid
A logical indicating whether to return the ggplot with a grid. Default is FALSE. The grid can be helpful if you want/need to move items around.

directory
File directory in which to save the R file. Defualt location is the current working directory.

filename
Name of the file, must end in '.R'. Default name is 'diagram_code.R'.

Value

A message telling the user where the file is.

Examples

varlabels <- c("S","I","R")
varnames <- c("Susceptible","Infected","Recovered")  # optional
flows <- list(S_flows = c("-b*S*I"),
              I_flows = c("b*S*I","-g*I"),
              R_flows = c("g*I"))
varlocations <-  matrix(data = c("S", "", "R", "", "I", "" ),
                        nrow = 2, ncol = 3, byrow = TRUE)
model_list <- list(varlabels = varlabels, varnames = varnames,
flows = flows, varlocations = varlocations)
diagram_list <- prepare_diagram(model_list = model_list)

# generate R code from model_list
write_diagram(model_list = model_list)

#> Error in menu(c("Yes", "No"), "Specified file already exists. Do you want to overwrite?"): menu() cannot be used non-interactively

# generate R code from diagram_list
write_diagram(diagram_list = diagram_list)

#> Error in menu(c("Yes", "No"), "Specified file already exists. Do you want to overwrite?"): menu() cannot be used non-interactively

# generate R code from both
write_diagram(model_list = model_list, diagram_list = diagram_list)

#> Error in menu(c("Yes", "No"), "Specified file already exists. Do you want to overwrite?"): menu() cannot be used non-interactively